fix(chart): Fixes an issue with aria label name clashes for a11y.
Specific inputs like `aria-label-button-close` are not allowed based on
the aria specification. Only known aria attributes are allowed to be
used. We have renamed the inputs in question to their camelcased
version, but kept the old ones for compatibility for now.
With version 7.0. we sould remove the kebap-cased inputs.

Fixes #526
Fixes #79